requirements.txt
# Core dependencies
numpy>=1.21.0
opencv-python>=4.5.0
scipy>=1.7.0

# Configuration and data handling
toml>=0.10.0
tomlkit>=0.11.0
PyYAML>=6.0

# Async and networking
aiofiles>=0.8.0
pyzmq>=22.0.0
pyserial>=3.5

# Computer vision and tracking
lap>=0.4.0
cython-bbox>=0.1.3

# RKNN (Rockchip Neural Network) - for embedded AI inference
# Note: This may need to be installed separately depending on your platform
# rknnlite>=1.4.0

# Event handling
pyee>=8.2.0

# Configuration management
yacs>=0.1.8

# GStreamer Python bindings (for video processing)
# Note: These require GStreamer to be installed on the system
PyGObject>=3.40.0

# Development and testing (optional)
pytest>=6.0.0
pytest-asyncio>=0.18.0

# Logging and utilities
pathlib2>=2.3.0; python_version < "3.4"

# Note: The 'status' module appears to be a custom module
# Make sure it's available in your Python path or install it separately
